server/requirements.txt
fastmcp>=2.5.0
psycopg2-binary>=2.9.9
PyJWT>=2.8.0
bcrypt>=4.1.0
starlette>=0.41.0
uvicorn[standard]>=0.30.0
